feat(csharp/): change size type to int and add some basic apis.

+++ csharp/Exceptions.cs
@@ -6,7 +6,7 @@ class InvalidLayoutException:Exception{
     // public InvalidLayoutException(TensorLayout layout){
     //     _message = $"{layout.ToString()} is an invalid TensorLayout.";
     // }
-    public InvalidLayoutException(Span<ulong> shape){
+    public InvalidLayoutException(Span<int> shape){
         _message = $"Shape is an invalid TensorLayout.";
     }
 }
